feat: implementa sistema de avaliação e rating por racha

+++ backend/racha-plus-demo-api/src/main/java/com/rachaplusdemo/api/dto/JogadorResponseDto.java
@@ -4,6 +4,6 @@ public record JogadorResponseDto(
         Long id,
         String nome,
         String email,
-        int rating
+        Double rating
 ) {
 }
